Revert "fix(EMI-2133): fix collections signals tracking (#14661)"
This reverts commit b25f903cf7ccc97c7935578a8ed82db9f7f53334.
