+++ src/Apps/Artwork/Components/ArtworkArtistSeries/ArtistSeriesArtworkRail.tsx
@@ -71,9 +71,9 @@ export const ArtistSeriesArtworkRail: React.FC<Props> = ({ artwork }) => {
                   destination_page_owner_slug: artwork.slug,
                   horizontal_slide_position: index,
                   type: "thumbnail",
-                  signal_label: getSignalLabel({
-                    collectorSignals: artwork.collectorSignals ?? {},
-                  }),
+                  signal_label: artwork.collectorSignals
+                    ? getSignalLabel(artwork.collectorSignals)
+                    : "",
                   signal_bid_count:
                     artwork.collectorSignals?.auction?.bidCount ?? undefined,
                   signal_lot_watcher_count:
