+++ src/Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarCommercialButtons.tsx
@@ -124,9 +124,9 @@ export const ArtworkSidebarCommercialButtons: React.FC<ArtworkSidebarCommercialB
       context_owner_type: OwnerType.artwork,
       context_owner_slug: artwork.slug,
       context_owner_id: artwork.internalID,
-      signal_label: getSignalLabel({
-        collectorSignals: artwork.collectorSignals ?? {},
-      }),
+      signal_label: artwork.collectorSignals
+        ? getSignalLabel(artwork.collectorSignals)
+        : "",
     }
     trackEvent(event)
 
@@ -140,9 +140,9 @@ export const ArtworkSidebarCommercialButtons: React.FC<ArtworkSidebarCommercialB
       context_owner_id: artwork.internalID,
       context_owner_slug: artwork.slug,
       flow: "Partner offer",
-      signal_label: getSignalLabel({
-        collectorSignals: artwork.collectorSignals ?? {},
-      }),
+      signal_label: artwork.collectorSignals
+        ? getSignalLabel(artwork.collectorSignals)
+        : "",
     }
 
     trackEvent(event)
@@ -198,9 +198,9 @@ export const ArtworkSidebarCommercialButtons: React.FC<ArtworkSidebarCommercialB
       context_owner_id: artwork.internalID,
       context_owner_slug: artwork.slug,
       flow: "Buy now",
-      signal_label: getSignalLabel({
-        collectorSignals: artwork.collectorSignals ?? {},
-      }),
+      signal_label: artwork.collectorSignals
+        ? getSignalLabel(artwork.collectorSignals)
+        : "",
     }
 
     trackEvent(event)
