+++ src/Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarPartnerInfo.tsx
@@ -61,9 +61,7 @@ const ArtworkSidebarPartnerInfo: React.FC<ArtworkSidebarPartnerInfoProps> = ({
       context_owner_type: OwnerType.artwork,
       context_owner_slug: slug,
       context_owner_id: internalID,
-      signal_label: getSignalLabel({
-        collectorSignals: collectorSignals ?? {},
-      }),
+      signal_label: collectorSignals ? getSignalLabel(collectorSignals) : "",
     }
 
     trackEvent(event)
