+++ src/Apps/Collect/Routes/Collection/index.tsx
@@ -17,6 +17,7 @@ import {
   SharedArtworkFilterContextProps,
 } from "Components/ArtworkFilter/ArtworkFilterContext"
 import { MetaTags } from "Components/MetaTags"
+import { ArtworkGridContextProvider } from "Components/ArtworkGrid/ArtworkGridContext"
 import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
 import { initializeVariablesWithFilterState } from "Apps/Collect/collectRoutes"
 import { useRouter } from "System/Hooks/useRouter"
@@ -47,6 +48,14 @@ export const CollectionApp: React.FC<CollectionAppProps> = props => {
 
   const socialImage = headerImage
 
+  const HIDE_SIGNAL_SLUGS = [
+    "trending-now",
+    "curators-picks-emerging-artists",
+    "curators-picks-blue-chip-artists",
+  ]
+
+  const hideSignals = HIDE_SIGNAL_SLUGS.includes(collection.slug)
+
   return (
     <StaticContainer shouldUpdate={!!match.elements}>
       <Analytics contextPageOwnerId={context.contextPageOwnerId as string}>
@@ -67,80 +76,82 @@ export const CollectionApp: React.FC<CollectionAppProps> = props => {
           <CollectionsHubRailsQueryRenderer slug={slug} />
 
           <FrameWithRecentlyViewed>
-            {/* TODO: Figure out why rerenders trigger refetches here, requiring
+            <ArtworkGridContextProvider hideSignals={hideSignals}>
+              {/* TODO: Figure out why rerenders trigger refetches here, requiring
               the static container to freeze rendering during route transitions. */}
-            <SystemQueryRenderer<CollectionArtworksQuery>
-              query={graphql`
-                query CollectionArtworksQuery(
-                  $slug: String!
-                  $aggregations: [ArtworkAggregation]
-                  $input: FilterArtworksInput!
-                  $shouldFetchCounts: Boolean!
-                ) {
-                  marketingCollection(slug: $slug) {
-                    ...CollectionArtworksFilter_collection
-                      @arguments(input: $input)
+              <SystemQueryRenderer<CollectionArtworksQuery>
+                query={graphql`
+                  query CollectionArtworksQuery(
+                    $slug: String!
+                    $aggregations: [ArtworkAggregation]
+                    $input: FilterArtworksInput!
+                    $shouldFetchCounts: Boolean!
+                  ) {
+                    marketingCollection(slug: $slug) {
+                      ...CollectionArtworksFilter_collection
+                        @arguments(input: $input)
 
-                    artworksConnection(
-                      aggregations: $aggregations
-                      includeMediumFilterInAggregation: true
-                      first: 20
-                      sort: "-decayed_merch"
-                    ) {
-                      counts @include(if: $shouldFetchCounts) {
-                        followedArtists
-                      }
-                      aggregations {
-                        slice
-                        counts {
-                          value
-                          name
-                          count
+                      artworksConnection(
+                        aggregations: $aggregations
+                        includeMediumFilterInAggregation: true
+                        first: 20
+                        sort: "-decayed_merch"
+                      ) {
+                        counts @include(if: $shouldFetchCounts) {
+                          followedArtists
+                        }
+                        aggregations {
+                          slice
+                          counts {
+                            value
+                            name
+                            count
+                          }
                         }
                       }
                     }
                   }
-                }
-              `}
-              variables={{
-                ...initializeVariablesWithFilterState(match.params, match),
-                slug,
-              }}
-              placeholder={<ArtworkFilterPlaceholder pt={6} />}
-              render={({ error, props }) => {
-                if (error) {
-                  console.error(
-                    "[collection]: Error loading artwork grid",
-                    error
+                `}
+                variables={{
+                  ...initializeVariablesWithFilterState(match.params, match),
+                  slug,
+                }}
+                placeholder={<ArtworkFilterPlaceholder pt={6} />}
+                render={({ error, props }) => {
+                  if (error) {
+                    console.error(
+                      "[collection]: Error loading artwork grid",
+                      error
+                    )
+                    return null
+                  }
+
+                  if (!props || !props.marketingCollection) {
+                    return <ArtworkFilterPlaceholder pt={6} />
+                  }
+
+                  return (
+                    <>
+                      <Spacer y={6} />
+
+                      <CollectionArtworksFilterRefetchContainer
+                        collection={props.marketingCollection}
+                        aggregations={
+                          props.marketingCollection?.artworksConnection
+                            ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
+                        }
+                        counts={
+                          props.marketingCollection.artworksConnection
+                            ?.counts as Counts
+                        }
+                      />
+
+                      <RelatedCollectionsRailQueryRenderer slug={slug} />
+                    </>
                   )
-                  return null
-                }
-
-                if (!props || !props.marketingCollection) {
-                  return <ArtworkFilterPlaceholder pt={6} />
-                }
-
-                return (
-                  <>
-                    <Spacer y={6} />
-
-                    <CollectionArtworksFilterRefetchContainer
-                      collection={props.marketingCollection}
-                      aggregations={
-                        props.marketingCollection?.artworksConnection
-                          ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
-                      }
-                      counts={
-                        props.marketingCollection.artworksConnection
-                          ?.counts as Counts
-                      }
-                    />
-
-                    <RelatedCollectionsRailQueryRenderer slug={slug} />
-                  </>
-                )
-              }}
-            />
+                }}
+              />
+            </ArtworkGridContextProvider>
           </FrameWithRecentlyViewed>
         </>
       </Analytics>
