+++ src/Apps/Home/Components/HomeEmergingPicksArtworksRail.tsx
@@ -35,9 +35,7 @@ export const HomeEmergingPicksArtworksRail: React.FC<HomeEmergingPicksArtworksRa
   }
 
   return (
-    <ArtworkGridContextProvider
-      hideSignals={["CURATORS_PICK", "INCREASED_INTEREST"]}
-    >
+    <ArtworkGridContextProvider hideSignals>
       <Rail
         title="Curators’ Picks: Emerging"
         subTitle="The best works by rising talents on Artsy, all available now."
@@ -73,10 +71,9 @@ export const HomeEmergingPicksArtworksRail: React.FC<HomeEmergingPicksArtworksRa
                   destination_page_owner_id: artwork.internalID,
                   destination_page_owner_slug: artwork.slug,
                   type: "thumbnail",
-                  signal_label: getSignalLabel({
-                    collectorSignals: artwork.collectorSignals ?? {},
-                    hideSignals: ["CURATORS_PICK", "INCREASED_INTEREST"],
-                  }),
+                  signal_label: artwork.collectorSignals
+                    ? getSignalLabel(artwork.collectorSignals)
+                    : "",
                   signal_bid_count:
                     artwork.collectorSignals?.auction?.bidCount ?? undefined,
                   signal_lot_watcher_count:
