+++ src/Apps/Home/Components/HomeNewWorksFromGalleriesYouFollowRail.tsx
@@ -75,9 +75,9 @@ const HomeNewWorksFromGalleriesYouFollowRail: React.FC<HomeNewWorksFromGalleries
                 destination_page_owner_id: artwork.internalID,
                 destination_page_owner_slug: artwork.slug,
                 type: "thumbnail",
-                signal_label: getSignalLabel({
-                  collectorSignals: artwork.collectorSignals ?? {},
-                }),
+                signal_label: artwork.collectorSignals
+                  ? getSignalLabel(artwork.collectorSignals)
+                  : "",
                 signal_bid_count:
                   artwork.collectorSignals?.auction?.bidCount ?? undefined,
                 signal_lot_watcher_count:
