+++ src/Apps/Home/Components/HomeRecentlyViewedRail.tsx
@@ -54,9 +54,9 @@ const HomeRecentlyViewedRail: React.FC<HomeRecentlyViewedRailProps> = ({
                 destination_page_owner_slug: artwork.slug,
                 destination_page_owner_type: OwnerType.artwork,
                 type: "thumbnail",
-                signal_label: getSignalLabel({
-                  collectorSignals: artwork.collectorSignals ?? {},
-                }),
+                signal_label: artwork.collectorSignals
+                  ? getSignalLabel(artwork.collectorSignals)
+                  : "",
                 signal_bid_count:
                   artwork.collectorSignals?.auction?.bidCount ?? undefined,
                 signal_lot_watcher_count:
