+++ src/Components/Notifications/Hooks/useNotificationsTracking.ts
@@ -53,9 +53,7 @@ export const useNotificationsTracking = () => {
         context_owner_id: artworkID,
         context_owner_slug: artworkSlug,
         flow: "Partner Offer",
-        signal_label: getSignalLabel({
-          collectorSignals: collectorSignals ?? {},
-        }),
+        signal_label: collectorSignals ? getSignalLabel(collectorSignals) : "",
       }
 
       trackEvent(event)
